test setup with jest and mock browser api
